+++ cookbooks/base/metadata.rb
@@ -4,6 +4,5 @@ maintainer_email 'carles.ml.dev@gmail.com'
 license          'Apache 2.0'
 description      'Installs/Configures base'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.18.0'
-depends          'core'
+version          '0.19.0'
 depends          'sudo'